refactor(ai): remove dead event flow and rename GPTResponseParser to VisionResponseParser
Delete the unreachable background meal-analysis event path (MealImageUploadedEvent,
MealAnalysisEventHandler, AnalyzeMealImageByUrlCommand/Handler) and associated
event_bus registrations. Remove the dead `analyze_by_url()` public method and
`_legacy_analysis_prompt()` from VisionAIService/strategy.

Rename GPTResponseParser → VisionResponseParser with a backward-compat shim so
all existing callers resolve without changes. Fix port docstring (was "OpenAI Vision
API"; now "Vision AI provider (currently Gemini)").

No behavior changes — pure dead code removal and cosmetic rename.
