+++ plans/260616-1533-ai-rearchitecture-beverage-scan/plan.md
@@ -29,7 +29,7 @@ Three coordinated goals shipped in sequence:
 
 | Phase | Name | Status | Effort |
 |-------|------|--------|--------|
-| 1 | [Dead Code Cleanup](./phase-01-dead-code-cleanup.md) | Pending | 0.5d |
+| 1 | [Dead Code Cleanup](./phase-01-dead-code-cleanup.md) | Completed | 0.5d |
 | 2 | [AI Core Consolidation](./phase-02-ai-core-consolidation.md) | Pending | 2–3d |
 | 3 | [Beverage Scanning and Single-Source Persistence](./phase-03-beverage-scanning-and-single-source-persistence.md) | Pending | 3–5d |
 | 4 | [Docs and Observability](./phase-04-docs-and-observability.md) | Pending | 0.5d |
