+++ src/app/commands/meal/__init__.py
@@ -9,13 +9,11 @@ from .edit_meal_command import (
     FoodItemChange,
     CustomNutritionData,
 )
-from .analyze_meal_image_by_url_command import AnalyzeMealImageByUrlCommand
 from .scan_by_url_command import ScanByUrlCommand
 from .upload_meal_image_immediately_command import UploadMealImageImmediatelyCommand
 
 __all__ = [
     "UploadMealImageImmediatelyCommand",
-    "AnalyzeMealImageByUrlCommand",
     "ScanByUrlCommand",
     "EditMealCommand",
     "AddCustomIngredientCommand",
