+++ src/app/handlers/command_handlers/__init__.py
@@ -4,7 +4,6 @@ Each handler is in its own file for better maintainability.
 """
 
 from .add_custom_ingredient_command_handler import AddCustomIngredientCommandHandler
-from .analyze_meal_image_by_url_command_handler import AnalyzeMealImageByUrlHandler
 from .scan_by_url_command_handler import ScanByUrlCommandHandler
 from .complete_onboarding_command_handler import CompleteOnboardingCommandHandler
 
@@ -84,7 +83,6 @@ __all__ = [
     "CreateManualMealCommandHandler",
     "ParseMealTextHandler",
     "UploadMealImageImmediatelyHandler",
-    "AnalyzeMealImageByUrlHandler",
     "ScanByUrlCommandHandler",
     # Notification handlers
     "RegisterFcmTokenCommandHandler",
