--- src/app/handlers/command_handlers/analyze_meal_image_by_url_command_handler.py
@@ -1,202 +1,0 @@
-"""
-Handler for immediate meal image analysis using pre-uploaded image URL.
-"""
-
-import logging
-import time
-from typing import Any
-from uuid import uuid4
-
-from src.app.commands.meal import AnalyzeMealImageByUrlCommand
-from src.app.events.base import EventHandler, handles
-from src.app.services.cache_invalidation_service import CacheInvalidationService
-from src.domain.model.meal import Meal, MealImage, MealStatus
-from src.domain.model.meal_projection import MealProjection
-from src.domain.parsers.gpt_response_parser import GPTResponseParser
-from src.domain.ports.async_unit_of_work_port import AsyncUnitOfWorkPort
-from src.domain.ports.vision_ai_service_port import VisionAIServicePort
-from src.domain.services.meal_analysis.deepl_meal_translation_service import (
-    DeepLMealTranslationService,
-)
-from src.domain.services.meal_type_determination_service import (
-    determine_meal_type_from_timestamp,
-)
-from src.domain.utils.timezone_utils import (
-    get_zone_info,
-    is_valid_timezone,
-    noon_utc_for_date,
-    utc_now,
-)
-
-logger = logging.getLogger(__name__)
-
-
-@handles(AnalyzeMealImageByUrlCommand)
-class AnalyzeMealImageByUrlHandler(EventHandler[AnalyzeMealImageByUrlCommand, Meal]):
-    """Handler for immediate meal image analysis from URL."""
-
-    def __init__(
-        self,
-        uow: AsyncUnitOfWorkPort,
-        event_bus: Any,
-        vision_service: VisionAIServicePort = None,
-        gpt_parser: GPTResponseParser = None,
-        meal_translation_service: DeepLMealTranslationService | None = None,
-        cache_invalidation: CacheInvalidationService | None = None,
-    ):
-        self.uow = uow
-        self.event_bus = event_bus
-        self.cache_invalidation = cache_invalidation
-        self.vision_service = vision_service
-        self.gpt_parser = gpt_parser
-        self.meal_translation_service = meal_translation_service
-
-    async def handle(self, command: AnalyzeMealImageByUrlCommand) -> Meal:
-        """Handle immediate meal image analysis from Cloudinary URL."""
-        if not all([self.vision_service, self.gpt_parser]):
-            raise RuntimeError("Required dependencies not configured")
-
-        try:
-            image_id = command.public_id.split("/")[-1]
-
-            async with self.uow as uow:
-                user_timezone = await uow.users.get_user_timezone(command.user_id)
-
-            if not user_timezone or not is_valid_timezone(user_timezone):
-                user_timezone = "UTC"
-                logger.info("Using UTC fallback for meal type detection")
-
-            now = utc_now()
-            meal_date = command.target_date if command.target_date else now.date()
-            if command.target_date and command.target_date != now.date():
-                # Past/future date — use noon to avoid date-boundary issues
-                meal_datetime = noon_utc_for_date(meal_date, user_timezone)
-                logger.info(f"Using noon for past/future date: {meal_datetime}")
-            else:
-                # Today or no date — use actual current time
-                meal_datetime = now
-                logger.info(f"Using current time: {meal_datetime}")
-
-            zone_info = get_zone_info(user_timezone)
-            local_datetime = meal_datetime.astimezone(zone_info)
-            meal_type = determine_meal_type_from_timestamp(local_datetime)
-
-            meal = Meal(
-                meal_id=str(uuid4()),
-                user_id=command.user_id,
-                status=MealStatus.ANALYZING,
-                created_at=meal_datetime,
-                meal_type=meal_type,
-                image=MealImage(
-                    image_id=image_id,
-                    format="jpeg" if "jpeg" in command.content_type else "png",
-                    size_bytes=command.file_size_bytes,
-                    url=command.image_url,
-                ),
-                source="scanner",
-            )
-
-            async with self.uow as uow:
-                saved_meal = await uow.meals.save(meal)
-                await uow.commit()
-
-            phase1_start = time.time()
-            if command.user_description:
-                from src.domain.strategies.meal_analysis_strategy import (
-                    AnalysisStrategyFactory,
-                )
-
-                strategy = AnalysisStrategyFactory.create_user_context_strategy(
-                    command.user_description
-                )
-                vision_result = await self.vision_service.analyze_by_url_with_strategy(
-                    command.image_url, strategy
-                )
-            else:
-                vision_result = await self.vision_service.analyze_by_url(
-                    command.image_url
-                )
-            phase1_elapsed = time.time() - phase1_start
-
-            if not self.gpt_parser.parse_is_food(vision_result):
-                raise ValueError(
-                    "Image does not appear to contain food. "
-                    "Please take a photo of food and try again."
-                )
-
-            nutrition = self.gpt_parser.parse_to_nutrition(vision_result)
-            dish_name = self.gpt_parser.parse_dish_name(vision_result)
-
-            has_food = (
-                nutrition
-                and nutrition.food_items
-                and len(nutrition.food_items) > 0
-                and nutrition.calories > 0
-            )
-            if not has_food:
-                raise ValueError(
-                    "No edible food detected in the image. "
-                    "Please take a photo of food and try again."
-                )
-
-            meal = meal.mark_ready(
-                nutrition=nutrition,
-                dish_name=dish_name or "Unknown dish",
-                raw_gpt_json=self.gpt_parser.extract_raw_json(vision_result),
-                emoji=self.gpt_parser.parse_emoji(vision_result),
-            )
-
-            phase2_elapsed = 0.0
-            if (
-                command.language
-                and command.language != "en"
-                and self.meal_translation_service
-            ):
-                phase2_start = time.time()
-                if nutrition and nutrition.food_items:
-                    try:
-                        await self.meal_translation_service.translate_meal(
-                            meal=saved_meal,
-                            dish_name=meal.dish_name,
-                            food_items=nutrition.food_items,
-                            target_language=command.language,
-                        )
-                    except Exception as e:
-                        logger.warning(
-                            "Translation failed for meal=%s: %s",
-                            saved_meal.meal_id,
-                            e,
-                        )
-                phase2_elapsed = time.time() - phase2_start
-
-            async with self.uow as uow:
-                final_meal = await uow.meals.save(meal)
-                await uow.commit()
-                logger.info(
-                    "[ANALYSIS-COMPLETE] meal=%s | total_elapsed=%.2fs | phase1=%.2fs | phase2=%.2fs | language=%s | status=%s",
-                    final_meal.meal_id,
-                    phase1_elapsed + phase2_elapsed,
-                    phase1_elapsed,
-                    phase2_elapsed,
-                    command.language,
-                    final_meal.status,
-                )
-
-            async with self.uow as uow:
-                final_meal = await uow.meals.find_by_id(
-                    meal.meal_id, projection=MealProjection.FULL_WITH_TRANSLATIONS
-                )
-
-            if self.cache_invalidation:
-                await self.cache_invalidation.after_meal_write(
-                    command.user_id, meal_date
-                )
-            return final_meal
-        except Exception as e:
-            logger.warning("Failed to analyze meal by URL: %s", type(e).__name__)
-            if "meal" in locals() and meal.meal_id:
-                meal = meal.mark_failed(error_message=str(e))
-                async with self.uow as uow:
-                    await uow.meals.save(meal)
-                    await uow.commit()
-            raise
