--- src/app/handlers/event_handlers/meal_analysis_event_handler.py
@@ -1,166 +1,0 @@
-"""
-Event handler for meal analysis events.
-"""
-
-import logging
-
-from src.app.events.base import EventHandler, handles
-from src.app.events.meal import MealImageUploadedEvent
-from src.domain.exceptions.ai_exceptions import (
-    AIOutputValidationError,
-    AIUnavailableError,
-)
-from src.domain.model.meal import MealStatus
-from src.domain.parsers.gpt_response_parser import GPTResponseParser
-from src.domain.ports.async_unit_of_work_port import AsyncUnitOfWorkPort
-from src.domain.ports.image_store_port import ImageStorePort
-from src.domain.ports.vision_ai_service_port import VisionAIServicePort
-from src.domain.services.meal_analysis.deepl_meal_translation_service import (
-    DeepLMealTranslationService,
-)
-from src.infra.adapters.cloudinary_image_store import CloudinaryImageStore
-from src.infra.adapters.vision_ai_service import VisionAIService
-from src.infra.database.uow_async import AsyncUnitOfWork
-
-logger = logging.getLogger(__name__)
-
-
-@handles(MealImageUploadedEvent)
-class MealAnalysisEventHandler(EventHandler[MealImageUploadedEvent, None]):
-    """Handler for meal analysis events."""
-
-    def __init__(
-        self,
-        uow: AsyncUnitOfWorkPort | None = None,
-        vision_service: VisionAIServicePort = None,
-        gpt_parser: GPTResponseParser = None,
-        image_store: ImageStorePort = None,
-        meal_translation_service: DeepLMealTranslationService | None = None,
-    ):
-        self.uow = uow
-        self.vision_service = vision_service or VisionAIService()
-        self.gpt_parser = gpt_parser or GPTResponseParser(strict_schema_mode=True)
-        self.image_store = image_store or CloudinaryImageStore()
-        self.meal_translation_service = meal_translation_service
-
-    def set_dependencies(self, **kwargs):
-        """Set dependencies for dependency injection."""
-        self.uow = kwargs.get("uow", self.uow)
-        self.vision_service = kwargs.get("vision_service", self.vision_service)
-        self.gpt_parser = kwargs.get("gpt_parser", self.gpt_parser)
-        self.image_store = kwargs.get("image_store", self.image_store)
-        self.meal_translation_service = kwargs.get(
-            "meal_translation_service", self.meal_translation_service
-        )
-
-    async def handle(self, event: MealImageUploadedEvent) -> None:
-        """Handle meal image uploaded event by triggering background analysis."""
-        logger.info(
-            f"EVENT HANDLER CALLED: Received MealImageUploadedEvent for meal {event.meal_id}"
-        )
-
-        # Phase 1: quick DB read — fetch meal and update status to ANALYZING.
-        # Connection is released before any slow external calls.
-        uow = self.uow or AsyncUnitOfWork()
-        async with uow:
-            meal = await uow.meals.find_by_id(event.meal_id)
-            if not meal:
-                logger.error(f"Meal {event.meal_id} not found for background analysis")
-                return
-            if meal.status != MealStatus.PROCESSING:
-                logger.info(
-                    f"Meal {event.meal_id} already processed with status {meal.status}"
-                )
-                return
-            meal = meal.mark_analyzing()
-            await uow.meals.save(meal)
-        # DB connection released here — vision and translation run without holding a connection
-
-        logger.info(f"Starting background analysis for meal {event.meal_id}")
-        try:
-            await self._perform_analysis(meal, language=event.language)
-        except Exception as e:
-            logger.error(
-                f"Error processing meal image upload event for meal {event.meal_id}: {str(e)}"
-            )
-            await self._mark_meal_as_failed(event.meal_id, str(e))
-
-    async def _perform_analysis(self, meal, language: str = "en"):
-        """Perform AI analysis. No DB connection is held during external calls."""
-        import asyncio
-
-        try:
-            await asyncio.sleep(1)
-
-            logger.info(f"Loading image contents for meal {meal.meal_id}")
-            image_contents = await self.image_store.load_async(meal.image.image_id)
-            if not image_contents:
-                raise Exception(
-                    f"Could not load image contents for image_id: {meal.image.image_id}"
-                )
-
-            logger.info(f"Performing real AI analysis for meal {meal.meal_id}")
-            vision_result = await self.vision_service.analyze(image_contents)
-
-            if not self.gpt_parser.parse_is_food(vision_result):
-                raise ValueError(
-                    "Image does not appear to contain food. "
-                    "Please take a photo of food and try again."
-                )
-
-            nutrition = self.gpt_parser.parse_to_nutrition(vision_result)
-            dish_name = self.gpt_parser.parse_dish_name(vision_result)
-            meal = meal.mark_ready(
-                nutrition=nutrition,
-                dish_name=dish_name or "Unknown dish",
-                raw_gpt_json=self.gpt_parser.extract_raw_json(vision_result),
-                emoji=self.gpt_parser.parse_emoji(vision_result),
-            )
-
-            if (
-                language
-                and language != "en"
-                and self.meal_translation_service
-                and nutrition
-                and nutrition.food_items
-            ):
-                try:
-                    await self.meal_translation_service.translate_meal(
-                        meal=meal,
-                        dish_name=meal.dish_name,
-                        food_items=nutrition.food_items,
-                        target_language=language,
-                    )
-                except Exception as e:
-                    logger.warning(f"Translation failed for meal {meal.meal_id}: {e}")
-
-            # Phase 3: quick DB write — save result (connection held only for this save)
-            uow = self.uow or AsyncUnitOfWork()
-            async with uow:
-                await uow.meals.save(meal)
-            logger.info(f"Real analysis completed for meal {meal.meal_id}")
-
-        except AIOutputValidationError:
-            logger.warning(
-                "[AI-OUTPUT-VALIDATION-EXHAUSTED] meal=%s purpose=meal_scan",
-                meal.meal_id,
-            )
-            await self._mark_meal_as_failed(
-                meal.meal_id,
-                "AI could not produce valid nutrition analysis. Please try again with a clearer photo.",
-            )
-        except AIUnavailableError as e:
-            logger.error("AI unavailable for meal %s: %s", meal.meal_id, e)
-            await self._mark_meal_as_failed(meal.meal_id, "AI service temporarily unavailable. Please try again.")
-        except Exception as e:
-            logger.error(f"Analysis failed for meal {meal.meal_id}: {str(e)}")
-            await self._mark_meal_as_failed(meal.meal_id, str(e))
-
-    async def _mark_meal_as_failed(self, meal_id: str, error_message: str):
-        """Mark a meal as failed with error message."""
-        uow = self.uow or AsyncUnitOfWork()
-        async with uow:
-            meal = await uow.meals.find_by_id(meal_id)
-            if meal:
-                await uow.meals.save(meal.mark_failed(error_message=error_message))
-                logger.info(f"Marked meal {meal_id} as failed")
