--- src/infra/services/ai/prompts/__init__.py
@@ -1,5 +1,0 @@
-"""AI prompt management."""
-
-from src.domain.services.prompts.system_prompts import SystemPrompts
-
-__all__ = ["SystemPrompts"]
