+++ tests/unit/api/test_event_bus_dependency_singletons.py
@@ -147,24 +147,18 @@ def test_get_configured_event_bus_is_singleton(monkeypatch):
         "SaveSuggestionCommandHandler",
         "DeleteSavedSuggestionCommandHandler",
         "GetSavedSuggestionsQueryHandler",
-        "MealAnalysisEventHandler",
     ]
-
-    class _StubWithHandle:
-        def handle(self, *a, **k):
-            pass
 
     for name in handler_names:
         if hasattr(mod, name):
-            stub = _StubWithHandle if name == "MealAnalysisEventHandler" else object
-            monkeypatch.setattr(mod, name, lambda *a, _stub=stub, **k: _stub())
+            monkeypatch.setattr(mod, name, lambda *a, **k: object())
 
     bus1 = mod.get_configured_event_bus()
     bus2 = mod.get_configured_event_bus()
     assert bus1 is bus2
     # Sanity: lots of registrations should have happened.
     assert len(bus1.handlers) > 10
-    assert len(bus1.subscriptions) == 1
+    assert len(bus1.subscriptions) == 0
 
 
 @pytest.mark.asyncio
@@ -216,7 +210,6 @@ async def test_configured_event_bus_can_send_movement_catalog_query(monkeypatch)
 
     handler_names = [
         "UploadMealImageImmediatelyHandler",
-        "AnalyzeMealImageByUrlHandler",
         "EditMealCommandHandler",
         "AddCustomIngredientCommandHandler",
         "DeleteMealCommandHandler",
