+++ tests/unit/handlers/command_handlers/test_ai_output_validation_error_propagation.py
@@ -91,51 +91,3 @@ async def test_scan_by_url_handler_propagates_ai_validation_error():
     assert exc_info.value is VALIDATION_ERROR
 
 
-@pytest.mark.asyncio
-async def test_background_event_handler_marks_meal_failed_on_validation_error():
-    """MealAnalysisEventHandler marks meal failed with sanitized message on AIOutputValidationError."""
-    from datetime import datetime
-
-    from src.app.handlers.event_handlers.meal_analysis_event_handler import (
-        MealAnalysisEventHandler,
-    )
-    from src.domain.model.meal import Meal, MealImage, MealStatus
-
-    meal = Meal(
-        meal_id="00000000-0000-0000-0000-000000000002",
-        user_id="00000000-0000-0000-0000-000000000001",
-        status=MealStatus.ANALYZING,
-        created_at=datetime.utcnow(),
-        image=MealImage(
-            image_id="00000000-0000-0000-0000-000000000003",
-            format="jpeg",
-            size_bytes=1000,
-            url="https://res.cloudinary.com/demo/image/upload/sample",
-        ),
-    )
-
-    mock_uow = MagicMock()
-    mock_uow.__aenter__ = AsyncMock(return_value=mock_uow)
-    mock_uow.__aexit__ = AsyncMock(return_value=False)
-    mock_uow.meals = MagicMock()
-    mock_uow.meals.find_by_id = AsyncMock(return_value=meal)
-    mock_uow.meals.save = AsyncMock(return_value=meal)
-
-    mock_image_store = MagicMock()
-    mock_image_store.load_async = AsyncMock(side_effect=VALIDATION_ERROR)
-
-    handler = MealAnalysisEventHandler(
-        uow=mock_uow,
-        vision_service=MagicMock(),
-        image_store=mock_image_store,
-    )
-
-    await handler._perform_analysis(meal, language="en")
-
-    mock_uow.meals.save.assert_awaited()
-    saved_meal = mock_uow.meals.save.call_args[0][0]
-    assert saved_meal.status == MealStatus.FAILED
-    error_msg = saved_meal.error_message
-    assert "validation_details" not in error_msg
-    assert "quantity_g" not in error_msg
-    assert len(error_msg) < 200
